refactor: UserSeat 모델 패키지 분리
